Minor fixes
Minor fixes

+++ src/SimpleWallet/SimpleWallet.cpp
@@ -321,8 +321,8 @@ struct TransferCommand {
           if (!remote_fee_address.empty()) {
             destination.address = remote_fee_address;
             int64_t remote_node_fee = static_cast<int64_t>(de.amount * 0.0025);
-            if (remote_node_fee > 100000000000)
-                remote_node_fee = 100000000000;
+            if (remote_node_fee > (int64_t)100000000000)
+                remote_node_fee = (int64_t)100000000000;
             destination.amount = remote_node_fee;
             dsts.push_back(destination);
           }
